requirements.txt
Django==2.0.5
django_tinymce==2.7.0
Pillow==5.4.1
beautifulsoup4==4.7.1
